FEAT: 작가노트 조회시 없는 경우 default 생성
