+++ src/main/java/com/there/src/artistStatement/ArtistStatementController.java
@@ -25,7 +25,6 @@ public class ArtistStatementController {
 
     final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private final ArtistStatementProvider artistStatementProvider;
     private final ArtistStatementService artistStatementService;
     private final JwtService jwtService;
 
@@ -44,7 +43,7 @@ public class ArtistStatementController {
     public BaseResponse<GetArtistStatementRes> getArtistStatement(@PathVariable("userIdx") int userIdx) {
         try {
 
-            GetArtistStatementRes getArtistStatementRes = artistStatementProvider.retrieveStatement(userIdx);
+            GetArtistStatementRes getArtistStatementRes = artistStatementService.retrieveStatement(userIdx);
             return new BaseResponse<>(getArtistStatementRes);
 
         } catch (BaseException exception) {
@@ -52,47 +51,47 @@ public class ArtistStatementController {
         }
     }
 
-    /**
-     * 작가노트 작성 API
-     * [POST] /statements/users/:userIdx
-     */
-    @ApiOperation(value="작가노트 작성 API", notes="유저의 작가노트를 작성합니다.")
-    @ApiResponses({
-            @ApiResponse(code = 1000, message = "요청에 성공하였습니다."),
-            @ApiResponse(code = 2003, message = "권한이 없는 유저의 접근입니다."),
-            @ApiResponse(code = 2110, message = "이미 작성된 작가노트가 있습니다."),
-            @ApiResponse(code = 2111, message = "자기소개 글자 수를 확인해주세요."),
-            @ApiResponse(code = 2112, message = "추구하는 작품 소개 글자 수를 확인해주세요."),
-            @ApiResponse(code = 2113, message = "연락처 글자 수를 확인해주세요."),
-            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다.")
-    })
-    @ResponseBody
-    @PostMapping("/users/{userIdx}")
-    public BaseResponse<PostArtistStatementRes> createArtistStatement(@PathVariable("userIdx") int userIdx,
-                                                                      @RequestBody PostArtistStatementReq postArtistStatementReq) {
-        try {
-
-            int userIdxByJwt = jwtService.getUserIdx1(jwtService.getJwt());
-            if (userIdxByJwt != userIdx) return new BaseResponse<>(INVALID_USER_JWT);
-
-            if(postArtistStatementReq.getSelfIntroduction().length() > 300) {
-                return new BaseResponse<>(STATEMENTS_INVALID_SELFINTRO);
-            }
-
-            if(postArtistStatementReq.getWorkIntroduction().length() > 300) {
-                return new BaseResponse<>(STATEMENTS_INVALID_WORKINTRO);
-            }
-
-            if(postArtistStatementReq.getContact().length() > 50) {
-                return new BaseResponse<>(STATEMENTS_INVALID_CONTACT);
-            }
-            PostArtistStatementRes postArtistStatementRes = artistStatementService.createStatement(userIdx, postArtistStatementReq);
-            return new BaseResponse<>(postArtistStatementRes);
-
-        } catch (BaseException exception) {
-            return new BaseResponse<>((exception.getStatus()));
-        }
-    }
+//    /**
+//     * 작가노트 작성 API
+//     * [POST] /statements/users/:userIdx
+//     */
+//    @ApiOperation(value="작가노트 작성 API", notes="유저의 작가노트를 작성합니다.")
+//    @ApiResponses({
+//            @ApiResponse(code = 1000, message = "요청에 성공하였습니다."),
+//            @ApiResponse(code = 2003, message = "권한이 없는 유저의 접근입니다."),
+//            @ApiResponse(code = 2110, message = "이미 작성된 작가노트가 있습니다."),
+//            @ApiResponse(code = 2111, message = "자기소개 글자 수를 확인해주세요."),
+//            @ApiResponse(code = 2112, message = "추구하는 작품 소개 글자 수를 확인해주세요."),
+//            @ApiResponse(code = 2113, message = "연락처 글자 수를 확인해주세요."),
+//            @ApiResponse(code = 4000, message = "데이터베이스 연결에 실패하였습니다.")
+//    })
+//    @ResponseBody
+//    @PostMapping("/users/{userIdx}")
+//    public BaseResponse<PostArtistStatementRes> createArtistStatement(@PathVariable("userIdx") int userIdx,
+//                                                                      @RequestBody PostArtistStatementReq postArtistStatementReq) {
+//        try {
+//
+//            int userIdxByJwt = jwtService.getUserIdx1(jwtService.getJwt());
+//            if (userIdxByJwt != userIdx) return new BaseResponse<>(INVALID_USER_JWT);
+//
+//            if(postArtistStatementReq.getSelfIntroduction().length() > 300) {
+//                return new BaseResponse<>(STATEMENTS_INVALID_SELFINTRO);
+//            }
+//
+//            if(postArtistStatementReq.getWorkIntroduction().length() > 300) {
+//                return new BaseResponse<>(STATEMENTS_INVALID_WORKINTRO);
+//            }
+//
+//            if(postArtistStatementReq.getContact().length() > 50) {
+//                return new BaseResponse<>(STATEMENTS_INVALID_CONTACT);
+//            }
+//            PostArtistStatementRes postArtistStatementRes = artistStatementService.createStatement(userIdx, postArtistStatementReq);
+//            return new BaseResponse<>(postArtistStatementRes);
+//
+//        } catch (BaseException exception) {
+//            return new BaseResponse<>((exception.getStatus()));
+//        }
+//    }
 
     /**
      * 작가노트 수정 API
